+++ web/streamlit/app.py
@@ -1,47 +1,35 @@
 import streamlit as st
+from transformers import AutoFeatureExtractor, ResNetForImageClassification
 import torch
-import torchvision.transforms as transforms
 from PIL import Image
-import torchvision.models as models
-import json
 
-# load labels
-with open('imagenet_labels.json') as f:
-    labels = json.load(f)
+# Load the pre-trained ResNet model and feature extractor
+feature_extractor = AutoFeatureExtractor.from_pretrained("microsoft/resnet-18")
+model = ResNetForImageClassification.from_pretrained("microsoft/resnet-18")
 
-def main():
-    st.title("PyTorch Image Classification")
-    # upload images
-    uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "jpeg", "png"])
-    if uploaded_file is not None:
-        # show images
-        image = Image.open(uploaded_file)
-        st.image(image, caption='Uploaded Image', use_column_width=True)
-        # classify
-        model = load_model()
-        prediction = classify_image(image, model)
-        st.write(f"Prediction: {prediction}")
-
-def load_model():
-    # load
-    model = models.resnet18(pretrained=True)
-    model.eval()
-    return model
-
-def classify_image(image, model):
-    # pre processing
-    transform = transforms.Compose([
-        transforms.Resize((224, 224)),
-        transforms.ToTensor(),
-        transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225))
-    ])
-    image = transform(image).unsqueeze(0)
-    # predict
+# Function to preprocess the image and make predictions
+def classify_image(image):
+    inputs = feature_extractor(image, return_tensors="pt")
     with torch.no_grad():
-        output = model(image)
-        _, predicted_idx = torch.max(output, 1)
-        predicted_label = labels[str(predicted_idx.item())]
-    return predicted_label
+        logits = model(**inputs).logits
+    predicted_label = logits.argmax(-1).item()
+    predicted_class = model.config.id2label[predicted_label]
+    return predicted_class
 
-if __name__ == '__main__':
-    main()
+# Streamlit app
+st.title("Image Classification with ResNet-18")
+
+# File uploader for user to upload their image
+uploaded_file = st.file_uploader("Choose an image...", type="jpg")
+
+# Display the uploaded image
+if uploaded_file is not None:
+    # Preprocess the image
+    image = Image.open(uploaded_file).convert("RGB")
+
+    # Make predictions
+    predicted_class = classify_image(image)
+
+    # Display the image and predictions
+    st.image(image, caption="Uploaded Image", use_column_width=True)
+    st.write(f"Predicted Label: {predicted_class}")
